% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{get_fold_data}
\alias{get_fold_data}
\title{Get the data associated with a fold}
\usage{
get_fold_data(species_name, data, folds, k, training)
}
\arguments{
\item{species_name}{Character vector. The name of the species you want get the fold data for.}

\item{data}{Dataframe. The occurrence or background data you want to get the fold data for.}

\item{folds}{Dataframe. The occurrence or background folds as created by \code{kfold_species_background}. Essentially a data.frame with a species column and the k-fold cross-validation logical vectors.}

\item{k}{Integer. Indicate which kth fold you want to return data for.}

\item{training}{Logical. If \code{TRUE} then training data is returned else if \code{FALSE} then test data is returned.}
}
\value{
Filtered version of data.
}
\description{
\code{get_fold_data} returns rows from data for a specific species for the kth fold
}
\examples{



}

