% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_occurrence_files}
\alias{get_occurrence_files}
\title{Get filenames of the occurrences files for some or all species}
\usage{
get_occurrence_file(species = NULL, raw = FALSE)
}
\arguments{
\item{species}{dataframe or character vector. Dataframe like returned by \code{\link{list_species}} or the names of the species. If \code{NULL} then all species occurrence files are returned.}

\item{raw}{logical. If \code{TRUE} the occurrence records without 25km² grid filtering and outlier removal. Default is \code{FALSE}.}
}
\description{
\code{get_occurrence_files} returns a character vector with the full paths to
the occurences files for all or a selection of species
}
\examples{
\dontrun{
all_paths <- get_occurrence_files()

species <- list_species()
first10 <- get_occurrence_files(species[1:10,])

abalistes_stellatus <- get_occurrence_files("Abalistes stellatus")
}
}

