% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_cv.R
\name{kfold_disc}
\alias{kfold_disc}
\title{Create k disc based cross-validation folds}
\usage{
kfold_disc(data, k = 5, distfun = NULL)
}
\description{
\code{kfold_disc} creates a k-fold partitioning of geographical data for cross-validation based on the distance between points. The n points nearest to a selected point are put into a group.
Returns a vector with fold numbers ranging from 1 to k.
}

