% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_background}
\alias{load_background}
\title{Load background data}
\usage{
load_background(type)
}
\arguments{
\item{type}{character. Either \code{"random"} or \code{"targetgroup"}.}
}
\value{
A dataframe with all background points.
}
\description{
\code{load_background} returns pre-generated background data.
}
\details{
The targetgroup background was created by subsampling an average of
  37 occurrence records (20000 in total) from each species in the dataset
  providing in essence the same sampling bias as the entire dataset.
}
\examples{
\dontrun{
random_bg <- load_background("random")
plot(random_bg[,2:3], pch=".", main="random background")

targetgroup_bg <- load_background("targetgroup")
plot(targetgroup_bg[,2:3], pch=".", main="targetgroup background")
}
}
\seealso{
\code{\link{get_fold_data}} \code{\link{get_occurrences}}
  \code{\link{load_folds}}
}

