% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_folds}
\alias{load_folds}
\title{Load folds}
\usage{
load_folds(fold_type = "pseudodisc")
}
\description{
Load folds
}
\details{
The different supported \code{fold_type} are:

\code{"pseudodisc"}: 5-fold disc partitioning of occurrences with pairwise
distance sampled and buffer filtered random background points, equivalent to
calling \code{\link{kfold_occurrence_background}} with
\code{occurrence_fold_type = "disc", k = 5, pwd_sample = TRUE,
background_buffer = 200*1000}

\code{"random"}:

\code{"targetgroup"}:
}

