% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marinespeed.R
\docType{package}
\name{marinespeed}
\alias{marinespeed}
\alias{marinespeed-package}
\title{marinespeed: Marine SPEcies and Environmental Data}
\description{
MarineSPEED is a benchmark dataset for marine presence-only species distribution modeling.
}

