% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismo.R
\name{pwdSample}
\alias{pwdSample}
\title{Pair-wise distance sampling}
\usage{
pwdSample(fixed, sample, reference, tr = 0.33, nearest = TRUE, n = 1,
  lonlat = TRUE, warn = TRUE)
}
\description{
For more information see \code{\link[dismo]{pwdSample}}
}

