% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{species_info}
\alias{species_info}
\title{Species information}
\usage{
species_info()
}
\description{
\code{species_info} returns a dataframe with species information.
}
\details{
Information about the taxonomy from the
  World Register of Marine Species (WoRMS), information on the covered
  latitudinal zones and the level of sample selection bias.
}
\examples{
info <- species_info()
unique(info$kingdom)
colnames(info)
}
\seealso{
\code{\link{list_species}}
}

