% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{species_traits}
\alias{species_traits}
\title{Species traits}
\usage{
species_traits(species = NULL)
}
\description{
\code{species_traits} returns a dataframe with species traits information.
}
\details{
Traits information includes information about the taxonomy from the World Register of
  Marine Species and habitat traits from the Encyclopedia of Life.
}
\examples{
traits <- species_traits()
traits$kingdom
}
\seealso{
\code{\link{list_species}}
}

