% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{species_traits}
\alias{species_traits}
\title{Species traits}
\usage{
species_traits()
}
\description{
\code{species_traits} returns a dataframe with species traits information.
}
\details{
Traits information includes information about the taxonomy from the
  World Register of Marine Species (WoRMS) and habitat traits from the
  Encyclopedia of Life (EOL).
}
\examples{
traits <- species_traits()
unique(traits$kingdom)
colnames(traits)
}
\seealso{
\code{\link{list_species}}
}

